requirements.txt
praw
python-dotenv
transformers
datasets
evaluate
accelerate
tabulate
yfinance
